switch ffi to uint32_t

+++ operator/halo2kzg/halo2kzg.go
@@ -42,11 +42,11 @@ func VerifyHalo2KzgProof(
 	publicInputPtr := (*C.uchar)(unsafe.Pointer(&publicInputBuffer[0]))
 	
 	// Cast data lengths to C-Types
-	proofLen := (C.size_t)(proofLen_u32)
-	csLen := (C.size_t)(csLen_u32) 
-	vkLen := (C.size_t)(vkLen_u32)
-	kzgParamLen := (C.size_t)(kzgParamLen_u32)
-	publicInputLen := (C.size_t)(publicInputLen_u32) 
+	proofLen := (C.uint32_t)(proofLen_u32)
+	csLen := (C.uint32_t)(csLen_u32) 
+	vkLen := (C.uint32_t)(vkLen_u32)
+	kzgParamLen := (C.uint32_t)(kzgParamLen_u32)
+	publicInputLen := (C.uint32_t)(publicInputLen_u32) 
 
 	return (bool)(C.verify_halo2_kzg_proof_ffi(
 		proofPtr, proofLen,
